+++ scout/adapter/mongo/event.py
@@ -155,6 +155,19 @@ class EventHandler(CaseEventHandler, VariantEventHandler):
 
         return self.event_collection.find(query).sort("created_at", pymongo.DESCENDING)
 
+    def case_events_by_verb(self, institute, case, verb):
+        """Return events with a specific verb for a case of an institute
+        Args:
+          institute (dict): an institute id
+          case (dict): a case id
+          verb (dict): an event action verb, example: "dismiss_variant"
+
+        Returns:
+            pymongo.Cursor: Query results
+        """
+        query = {"institute": institute["_id"], "case": case["_id"], "verb": verb}
+        return self.event_collection.find(query)
+
     def user_events(self, user_obj=None):
         """Fetch all events by a specific user."""
         query = dict(user_id=user_obj["_id"]) if user_obj else dict()