+++ CHANGELOG.md
@@ -7,6 +7,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [x.x.x]
 ### Fixed
 ### Added
+- Gray background for dismissed compounds in variants list and variant page
+- Pin badge for pinned compounds in variants list and variant page
 ### Changed
 
 ## [4.33]
@@ -20,8 +22,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Autodeploy docs on release
 - Documentation for updating case individuals tracks
 - Filter cases and dashboard stats by analysis track
-- Gray background for dismissed compounds in variants list and variant page
-- Pin badge for pinned compounds in variants list and variant page
 ### Changed
 - Changed from deprecated db update method
 - Pre-selected fields to run queries with in dashboard page